Scaffolding for editor
